refactor(native): isolate gcc-like toolchain policy
